api/djangormm/requirements.txt
amqp==2.5.1
appdirs==1.4.3
asn1crypto==1.0.1
attrs==19.2.0
billiard==3.6.1.0
black==19.3b0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
distro==1.4.0
Django==2.2.6
django-cors-headers==3.1.1
django-rest-knox==4.1.0
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.5
loguru==0.3.2
more-itertools==7.2.0
psycopg2-binary==2.8.3
pycparser==2.19
pymongo==3.9.0
pyotp==2.3.0
pytz==2019.2
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.6
uWSGI==2.0.18
validators==0.14.0
vine==1.3.0
zipp==0.6.0
